fix(contracts): validate scheduled transaction timestamp during execution rather than  validation
block.timestamp isn't available during verification, see https://docs.zksync.io/build/developer-reference/account-abstraction.html#what-could-be-allowed-in-the-future

+++ contracts/contracts/policy/hooks/DelayHook.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.20;
 import {Scheduler} from '../../libraries/Scheduler.sol';
 
 struct DelayConfig {
-  uint256 delay /* seconds */;
+  uint32 delay /* seconds */;
 }
 
 library DelayHook {
@@ -15,7 +15,7 @@ library DelayHook {
     DelayConfig memory config = abi.decode(configData, (DelayConfig));
     if (config.delay == 0) return true;
 
-    Scheduler.schedule(proposal, block.timestamp + config.delay);
+    Scheduler.schedule(proposal, uint32(block.timestamp) + config.delay);
     return false;
   }
 }
